+++ contracts/src/token/erc20/extensions/erc4626.rs
@@ -9,11 +9,10 @@
 
 use alloc::{vec, vec::Vec};
 
-use alloy_primitives::{uint, Address, U256, U8};
-use alloy_sol_types::SolCall;
+use alloy_primitives::{uint, Address, FixedBytes, U256, U8};
 pub use sol::*;
 use stylus_sdk::{
-    call::{self, Call, MethodError},
+    call::{Call, MethodError},
     contract, evm, msg,
     prelude::*,
     storage::{StorageAddress, StorageU8},
@@ -22,13 +21,14 @@ use stylus_sdk::{
 use crate::{
     token::erc20::{
         self,
-        utils::{
-            safe_erc20, IERC20Metadata, IErc20 as IErc20Solidity, ISafeErc20,
-            SafeErc20,
-        },
+        interface::{Erc20Interface, IERC20Metadata},
+        utils::{safe_erc20, ISafeErc20, SafeErc20},
         Erc20, IErc20,
     },
-    utils::math::alloy::{Math, Rounding},
+    utils::{
+        introspection::erc165::{Erc165, IErc165},
+        math::alloy::{Math, Rounding},
+    },
 };
 
 const ONE: U256 = uint!(1_U256);
@@ -40,10 +40,12 @@ mod sol {
 
     sol! {
         /// Emitted when assets are deposited into the contract.
+        #[derive(Debug)]
         #[allow(missing_docs)]
         event Deposit(address indexed sender, address indexed owner, uint256 assets, uint256 shares);
 
         /// Emitted when assets are withdrawn from the contract.
+        #[derive(Debug)]
         #[allow(missing_docs)]
         event Withdraw(
             address indexed sender,
@@ -155,6 +157,65 @@ unsafe impl TopLevelStorage for Erc4626 {}
 pub trait IErc4626 {
     /// The error type associated to the trait implementation.
     type Error: Into<alloc::vec::Vec<u8>>;
+
+    // Manually calculated, as some of the functions' parameters do not
+    // implement AbiType.
+    /// Solidity interface id associated with [`IErc4626`] trait. Computed as a
+    /// XOR of selectors for each function in the trait.
+    const INTERFACE_ID: u32 =
+        u32::from_be_bytes(stylus_sdk::function_selector!("asset"))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!("totalAssets"))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "convertToShares",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "convertToAssets",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "maxDeposit",
+                Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "previewDeposit",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "deposit", U256, Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "maxMint", Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "previewMint",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "mint", U256, Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "maxWithdraw",
+                Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "previewWithdraw",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "withdraw", U256, Address, Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "maxRedeem",
+                Address
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "previewRedeem",
+                U256
+            ))
+            ^ u32::from_be_bytes(stylus_sdk::function_selector!(
+                "redeem", U256, Address, Address
+            ));
 
     /// Returns the address of the underlying token used for the Vault for
     /// accounting, depositing, and withdrawing.
@@ -753,11 +814,11 @@ impl IErc4626 for Erc4626 {
     type Error = Error;
 
     fn asset(&self) -> Address {
-        *self.asset
+        self.asset.get()
     }
 
     fn total_assets(&mut self) -> Result<U256, Self::Error> {
-        let erc20 = IErc20Solidity::new(self.asset());
+        let erc20 = Erc20Interface::new(self.asset());
         let call = Call::new_in(self);
         Ok(erc20
             .balance_of(call, contract::address())
@@ -942,14 +1003,10 @@ impl Erc4626 {
     /// in any way. This follows Rust's idiomatic Option pattern rather than
     /// Solidity's boolean tuple return.
     fn try_get_asset_decimals(&mut self, asset: Address) -> Option<u8> {
-        let call = IERC20Metadata::decimalsCall {};
-
-        // Perform the static call
-        let decimals =
-            call::static_call(Call::new_in(self), asset, &call.abi_encode())
-                .ok()?;
-
-        decimals.first().copied()
+        let erc20 = IERC20Metadata::new(asset);
+        let call = Call::new_in(self);
+        let decimals = erc20.decimals(call).ok();
+        decimals
     }
 }
 
@@ -1184,15 +1241,22 @@ impl Erc4626 {
     }
 }
 
+impl IErc165 for Erc4626 {
+    fn supports_interface(interface_id: FixedBytes<4>) -> bool {
+        <Self as IErc4626>::INTERFACE_ID == u32::from_be_bytes(*interface_id)
+            || Erc165::supports_interface(interface_id)
+    }
+}
+
 // TODO: Add missing tests once `motsu` supports calling external contracts.
 #[cfg(all(test, feature = "std"))]
 mod tests {
     use alloy_primitives::{address, Address, U256, U8};
-    use motsu::prelude::Contract;
+    use motsu::prelude::*;
     use stylus_sdk::prelude::*;
 
     use super::{Erc4626, IErc4626};
-    use crate::token::erc20::Erc20;
+    use crate::{token::erc20::Erc20, utils::introspection::erc165::IErc165};
 
     #[storage]
     struct Erc4626TestExample {
@@ -1245,7 +1309,10 @@ mod tests {
     ) {
         let assets = U256::from(1000);
         contract.init(alice, |contract| {
-            contract.erc20._mint(alice, assets).expect("should mint assets");
+            contract
+                .erc20
+                ._mint(alice, assets)
+                .motsu_expect("should mint assets");
         });
         let max_redeem = contract.sender(alice).max_redeem(alice);
         assert_eq!(assets, max_redeem);
@@ -1277,5 +1344,25 @@ mod tests {
 
         let decimals = contract.sender(alice).erc4626.decimals();
         assert_eq!(decimals, underlying_decimals + new_decimal_offset);
+    }
+
+    #[motsu::test]
+    fn interface_id() {
+        let actual = <Erc4626 as IErc4626>::INTERFACE_ID;
+        let expected = 0x87dfe5a0;
+        assert_eq!(actual, expected);
+    }
+
+    #[motsu::test]
+    fn supports_interface() {
+        assert!(Erc4626::supports_interface(
+            <Erc4626 as IErc4626>::INTERFACE_ID.into()
+        ));
+        assert!(Erc4626::supports_interface(
+            <Erc4626 as IErc165>::INTERFACE_ID.into()
+        ));
+
+        let fake_interface_id = 0x12345678u32;
+        assert!(!Erc4626::supports_interface(fake_interface_id.into()));
     }
 }