+++ contracts/src/token/erc20/interface.rs
@@ -1,0 +1,30 @@
+//! Solidity Interface of the ERC-20 token.
+pub use token::*;
+
+mod token {
+    #![allow(missing_docs)]
+    #![cfg_attr(coverage_nightly, coverage(off))]
+
+    use alloc::vec;
+
+    use stylus_sdk::prelude::sol_interface;
+    sol_interface! {
+        interface Erc20Interface {
+            function totalSupply() external view returns (uint256);
+            function balanceOf(address account) external view returns (uint256);
+            function transfer(address to, uint256 value) external returns (bool);
+            function allowance(address owner, address spender) external view returns (uint256);
+            function approve(address spender, uint256 value) external returns (bool);
+            function transferFrom(address from, address to, uint256 value) external returns (bool);
+        }
+    }
+
+    sol_interface! {
+        /// Solidity Interface of the ERC-20 Metadata token.
+        interface IERC20Metadata {
+            function name() external view returns (string memory);
+            function symbol() external view returns (string memory);
+            function decimals() external view returns (uint8);
+        }
+    }
+}