+++ CHANGELOG.md
@@ -9,9 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Erc721Wrapper` extension to support token wrapping. #461
+- Add callable interface for ERC-721. #461
+- Add missing functions to callable ERC-20 interface. #461
+- All events now derive `Debug`. #614
 - `Erc20Wrapper` extension to support token wrapping. #498
 - `Erc20` events derive `Debug`. #498
 - Implement `MethodError` for all contracts' errors. #594
+- Implement `IErc165` for all base contracts for standard interface detection. #603
+- Expose `INTERFACE_ID` for `Erc20Wrapper`, `Erc4626` and `Erc20FlashMint`. #603
+- Short Weierstrass elliptic curves primitives. #589
 
 ### Changed
 
@@ -19,11 +26,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed (Breaking)
 
+- Moved `Erc20` callable interface to _/erc20/interface.rs_ module and renamed it to `Erc20Interface`. #461
+- Bump `cargo-stylus` to `v0.5.11`. #617
+- Bump Stylus SDK to `v0.8.4`. #624
 - Remove `ownable_two_step::Error` wrapper in `Ownable2Step`, and emit `ownable::Error` directly. #594
+- Poseidon babybear and goldilocks (64-bit) instances now have 256-bit security (capacity 4). #613
+- `BitIteratorBE` (bit iteration) trait at `openzeppelin_crypto` now accepts `self` by value. #589
 
 ### Fixed
 
--
+- `IErc20Metadata::INTERFACE_ID` now has the correct value.
 
 ## [v0.2.0-alpha.4] - 2025-03-06
 
@@ -38,7 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor `Erc20Permit` extension to be a composition of `Erc20` and `Nonces` contracts. #574
 - Replace `VestingWallet::receive_ether` with dedicated `receive` function. #529
 - Extract `IAccessControl` trait from `AccessControl` contract. #527
-- Bump Stylus SDK to v0.8.1 #587
+- Bump Stylus SDK to `v0.8.1` #587
 
 ### Fixed
 
@@ -65,7 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - All contract state fields are no longer public. #500
 - `Erc721Consecutive::_mint_consecutive` turned into an internal function. #500
-- Bump cargo-stylus to v0.5.8. #493
+- Bump `cargo-stylus` to `v0.5.8`. #493
 - Constants `TYPE_HASH`, `FIELDS`, `SALT` and `TYPED_DATA_PREFIX`, and type `DomainSeparatorTuple` are no longer exported from `utils::cryptography::eip712`. #478
 - Bump Stylus SDK to v0.7.0. #433
 - Bump `alloy` dependencies to v0.8.14. #433